# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Order Management API
  version: 1.0.0
  description: |
    Enterprise-grade Order Management API for B2B retail supply chain.
    
    Technical Highlights:
    - **API Versioning**: Global versioning via `/api/v1` prefix.
    - **Runtime**: Java 21 with **Virtual Threads** (Project Loom).
    - **Server**: Jetty.
    - **Observability**: Full-stack traceability via `X-Request-Id`.
    
    Calculation Rule:
    `totalCost = orderAmount * unitPriceSnapshot * (1 + taxRateSnapshot)`
    
    All timestamps follow **ISO 8601** format (e.g., 2026-04-16T22:30:00Z).

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Product
  - name: Order
  - name: User

paths:
  /api/v1/product/{product_id}:
    get:
      tags: [Product]
      summary: Get Product by Id
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product found successfully
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Product'}
        '404':
          description: Product not found
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}

  /api/v1/order:
    post:
      tags: [Order]
      summary: Create Order
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateOrderRequest'}
      responses:
        '201':
          description: Order created successfully
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
            Location: {description: URL, schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateOrderResponse'}
        '400':
          description: Invalid request
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: User or Product not found
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '409':
          description: Conflict with current resource state or request data
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '422':
          description: Idempotency key reused with different request payload
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}

  /api/v1/order/{order_id}:
    patch:
      tags: [Order]
      summary: Patch Order by order_id
      description: |
        Perform controlled partial updates (orderAmount only).
        Updates are subject to state-machine rules.
      operationId: patchOrderById
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PatchOrderRequest'}
      responses:
        '200':
          description: Order updated
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Order'}
        '400':
          description: Business rule violation
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Order not found
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '409':
          description: Conflict with current resource state or request data
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    delete:
      tags: [Order]
      summary: Delete Order by order_id
      operationId: deleteOrderById
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '204':
          description: Order deleted
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
        '404':
          description: Order not found
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}

  /api/v1/order/{userId}:
    get:
      tags: [Order]
      summary: Get Order by User
      description: |
        Retrieve active orders for a specific user with pagination.
        Default sorting: **created_at DESC, id DESC** to ensure stability and newest-first experience.
      operationId: getOrdersByUserId
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: page
          in: query
          description: |
            Page number (1-indexed). 
            Offset-based pagination is chosen to support random page access required by business operations.
          schema: {type: integer, default: 1, minimum: 1}
        - name: size
          in: query
          description: Number of records per page. Limited to 100 to prevent deep-page performance issues.
          schema: {type: integer, default: 20, minimum: 1, maximum: 100}
      responses:
        '200':
          description: Paged list of user orders
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PagedOrderResponse'}
        '404':
          description: User not found
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}

  /api/v1/user/{userId}:
    delete:
      tags: [User]
      summary: Delete User
      description: |
        Soft delete a user record.
        In compliance with the assignment, all related user order references are also soft-deleted 
        to maintain business record consistency while removing them from active operations.
      operationId: deleteUserById
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deleted
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
        '404':
          description: User not found
          headers:
            X-Request-Id: {$ref: '#/components/headers/X-Request-Id'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}

components:
  headers:
    X-Request-Id:
      description: Unique trace ID for correlation.
      schema: {type: string, format: uuid}
    Idempotency-Key:
      description: Unique key to ensure create-order requests are processed at most once for the same semantic request.
      schema: {type: string, minLength: 1, maxLength: 128}

  parameters:
    ProductId: {name: product_id, in: path, required: true, schema: {type: string, format: uuid}}
    OrderId: {name: order_id, in: path, required: true, schema: {type: string, format: uuid}}
    UserId: {name: userId, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key supplied by the client to deduplicate retried create-order requests.
      schema: {type: string, minLength: 1, maxLength: 128}

  schemas:
    Product:
      type: object
      required: [id, productName, productCategoryId, unitPrice, taxRate, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        productName: {type: string}
        productCategoryId: {type: string, format: uuid}
        unitPrice: {type: number}
        taxRate: {type: number}
        createdAt: {type: string, format: date-time, example: "2026-04-16T22:30:00Z"}
        updatedAt: {type: string, format: date-time, example: "2026-04-16T22:30:00Z"}

    Order:
      type: object
      required: [id, userId, productId, orderAmount, unitPrice, taxRate, totalCost, status, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        userId: {type: string, format: uuid}
        productId: {type: string, format: uuid}
        orderAmount: {type: integer}
        unitPrice: {type: number, minimum: 0, multipleOf: 0.0001, example: 0.0001}
        taxRate: {type: number, minimum: 0, multipleOf: 0.0001, example: 0.0001}
        totalCost: {type: number, minimum: 0, multipleOf: 0.0001, example: 0.0001}
        status: {$ref: '#/components/schemas/OrderStatus'}
        createdAt: {type: string, format: date-time, example: "2026-04-16T22:30:00Z"}
        updatedAt: {type: string, format: date-time, example: "2026-04-16T22:30:00Z"}

    OrderStatus:
      type: string
      enum: [DRAFT, SUBMITTED, APPROVED, CANCELLED]

    PagedOrderResponse:
      type: object
      required: [data, pagination]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Order'}}
        pagination: {$ref: '#/components/schemas/PaginationInfo'}

    PaginationInfo:
      type: object
      required: [page, size, totalItems, totalPages]
      properties:
        page: {type: integer, example: 1}
        size: {type: integer, example: 20}
        totalItems: {type: integer, format: int64, example: 100}
        totalPages: {type: integer, example: 5}

    CreateOrderRequest:
      type: object
      required: [userId, productId, orderAmount, expectedUnitPrice, expectedTaxRate]
      properties:
        userId: {type: string, format: uuid}
        productId: {type: string, format: uuid}
        orderAmount: {type: integer, minimum: 1, maximum: 2147483647}
        expectedUnitPrice: {type: number, minimum: 0, multipleOf: 0.0001, example: 0.0001}
        expectedTaxRate: {type: number, minimum: 0, multipleOf: 0.0001, example: 0.0001}

    CreateOrderResponse:
      type: object
      required: [id]
      properties:
        id: {type: string, format: uuid}

    PatchOrderRequest:
      type: object
      required: [orderAmount, expectedUnitPrice, expectedTaxRate]
      properties:
        orderAmount: {type: integer, minimum: 1, maximum: 2147483647}
        expectedUnitPrice: {type: number, minimum: 0, multipleOf: 0.0001, example: 0.0001}
        expectedTaxRate: {type: number, minimum: 0, multipleOf: 0.0001, example: 0.0001}

    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code: {type: string}
        message: {type: string}
        requestId: {type: string}
        details: {type: array, items: {$ref: '#/components/schemas/ErrorDetail'}}

    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field: {type: string}
        message: {type: string}